Add toolbar, title hairstyle screen, add book appointment screen

+++ app/src/main/java/com/example/barbershopmanagementapp/Hairstyles/HairstyleModel.java
@@ -26,7 +26,7 @@ public class HairstyleModel {
         this.imgId = imgId;
     }
 
-    public long getPrice() {
+    public int getPrice() {
         return price;
     }
